requirements.txt
jittor>=1.3.1.21
numpy
Pillow
six
python-lmdb
python-Levenshtein
opencv-python